# zend_compile: Optimize `sprintf()` into a rope