@TimWolla 2024-06-12
Regarding the “initial version” remark: Once this is reviewed and merged, I hope to also add support for plain `%d` placeholders, which I expect to be the second most common type of placeholder. I expect that to just require an additional `(int)` cast, given that `%d` just prints integers as-is.

“Named placeholders” (i.e. `%1$s`) might also be possible to support as a follow-up, given that they do not contain any additional logic either.