@alexandre-daubois 2024-06-13
Hi! This fail just popped in Symfony CI: https://github.com/symfony/symfony/actions/runs/9497428991/job/26173977979. It seems this PR triggered it (older CI runs don't have the failure).

```
1) Symfony\Component\VarDumper\Tests\Caster\FFICasterTest::testCastNonTrailingCharPointer
Failed asserting that string matches format description.
--- Expected
+++ Actual
@@ @@
 FFI\CData<char*> size 8 align 8 {
-  cdata: "Hello World!%s"
+  cdata: b"Hello World!\x011RK`\x1EÐþ\x16\x7F\x00"
 }

/home/runner/work/symfony/symfony/src/Symfony/Component/VarDumper/Test/VarDumperTestTrait.php:52
/home/runner/work/symfony/symfony/src/Symfony/Component/VarDumper/Tests/Caster/FFICasterTest.php:219
```

Could this be considered a BC break? It's just this test that's causing the problem, maybe it's a fix to be made on the Symfony side. Just thought I'd give you a heads-up in case it alerts you to something obvious. 🙂