@TimWolla 2024-06-13
@alexandre-daubois Thank you for the report. I've taken a look and it appears the issue is unrelated. It also happens when I disable the optimization by inserting `return FAILURE;` at the top of the `zend_compile_func_sprintf()` function. I am also not seeing any calls to `sprintf()` nearby the test (`assertDumpMatchesFormat()` does not use it).

Note that I am only able to reproduce the issue with a production build, a debug build of PHP (`--enable-debug`) does not reproduce it.