@TimWolla 2024-06-13
It's also visible in the CI log that the PHP 8.4 build does not yet include this PR:

> ✓ PHP Installed PHP 8.4.0-dev (18cfd94de402e1a843b316f15893f4a0d22a374e)

18cfd9 is the commit right before the merge of this PR.