@alexandre-daubois 2024-06-13
Oh right, that's weird. I'll have a deeper look at the test case, thank you for having a look at it!