@arnaud-lb 2024-06-13
I believe it's not a bug in php-src. I've reported symfony issue: https://github.com/symfony/symfony/issues/57387