@mvorisek 2024-06-18
@TimWolla I have two questions about this PR:

- can `vsprintf` be optimized as well for consistency?
- why is `sprintf` not marked as CTE [1] (compile time evaluable)?

[1] https://github.com/php/php-src/blob/e4a8d5b16f/ext/standard/basic_functions.stub.php#L2999