@iluuu1994 2024-06-18
@mvorisek Because they are locale-dependent for floating points.