@TimWolla 2024-06-18
> can `vsprintf` be optimized as well for consistency?

From my experience `vsprintf()` is used much more rarely compared to `sprintf()` and if it is used then the format string is not a string literal. Combining a format string literal with an array parameter seems like an odd choice, because you might as well pass the parameters separately.

For that reason I don't plan to do that, but feel free to send a PR if you have data that indicates it would be worth it.