@Girgias 2024-06-19
> @mvorisek Because they are locale-dependent for floating points.

That's incorrect since 8.0 with the [Locale-independent float to string cast](https://wiki.php.net/rfc/locale_independent_float_to_string) RFC got accepted in 8.0.
However, floating points still depend on the `'precision'` INI setting: https://3v4l.org/mMjuK